requirements.txt
python-magic-bin
eyeD3